WS-NA-add-chromatic-tests: create new file
